chore: :truck: rename api file
